+++ src/agent/samplers/hw_sensors/mod.rs
@@ -1,0 +1,7 @@
+#[cfg(target_os = "linux")]
+mod linux;
+
+#[cfg(not(target_os = "linux"))]
+mod stats {
+    include!("./linux/stats.rs");
+}